pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ssms-gui"
version = "0.1.0"
description = "GUI prototype"
authors = [
	{name = "George Dang", "email" = "53052793+gtdang@users.noreply.github.com"},
	{name =  "Alexander Fengler", email = "alexander_fengler@brown.edu"}
	]
readme = "README.md"
dependencies= ['ssm-simulators', 'streamlit>1.30.0', 'matplotlib', 'seaborn']
requires-python= ">=3.10, <3.12"

[project.optional-dependencies]
dev = ["pytest", "black", "mypy", "isort", "flake8", "pre-commit"]